api/requirements.txt
aio-pika==6.6.0
aiohttp==3.6.2
aiohttp-apispec==2.2.1
aiohttp-cors==0.7.0
aiohttp-jwt==0.5.0
aiohttp-session==2.9.0
aiohttp-swagger==1.0.14
aiormq==3.2.1
apispec==3.3.0
async-timeout==3.0.1
attrs==19.3.0
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
idna==2.9
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.5.1
motor==2.1.0
multidict==4.7.5
pamqp==2.3.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
PyYAML==5.3
six==1.14.0
webargs==5.5.3
yarl==1.4.2
Pillow